move music fragment
