+++ app/src/main/java/com/example/genshinwish/fragments/MusicFragment.kt
@@ -1,18 +1,29 @@
 package com.example.genshinwish.fragments
 
+import android.content.Intent
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import com.example.genshinwish.R
 import com.example.genshinwish.databinding.FragmentMusicBinding
-
+import com.example.genshinwish.models.Song
+import com.example.genshinwish.notification.Mp3Service
 
 class MusicFragment : Fragment() {
    private lateinit var binding : FragmentMusicBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        binding = DataBindingUtil.setContentView(requireActivity(), R.layout.fragment_music)
+        binding.btnStartService.setOnClickListener {
+            clickStartService()
+        }
+        binding.btnStopService.setOnClickListener {
+            clickStopService()
+        }
     }
 
     override fun onCreateView(
@@ -21,5 +32,21 @@ class MusicFragment : Fragment() {
     ): View? {
         binding = FragmentMusicBinding.inflate(inflater)
         return binding.root
+    }
+
+    private fun clickStopService() {
+        val intent = Intent(activity, Mp3Service::class.java)
+            requireActivity().stopService(intent)
+    }
+
+    private fun clickStartService() {
+        val intent = Intent(activity, Mp3Service::class.java)
+        val bundle = Bundle()
+        val song = Song("Genshin Impact Battle Song","Paimon",
+            R.drawable.razor,
+            R.raw.battle_paimon)
+        bundle.putSerializable("object_song",song)
+        intent.putExtra("Bundle",bundle)
+        requireActivity().startService(intent)
     }
 }
