+++ app/src/main/res/layout/activity_second.xml
@@ -1,57 +1,71 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
-<RelativeLayout
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".activities.SecondActivity">
 
-    <ImageView
-        android:id="@+id/img_song"
+    <FrameLayout
         android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:src="@drawable/hutao_banner"
-        />
-    <LinearLayout
-        android:id="@+id/song_infor"
-        android:layout_width="match_parent"
-        android:layout_height="100dp"
-        android:layout_below="@+id/img_song"
-        android:orientation="vertical">
+        android:layout_height="match_parent"
+        tools:context=".fragments.PublicFragment">
+
+        <!-- TODO: Update blank fragment layout -->
         <TextView
-            android:id="@+id/txt_song_name"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textAlignment="center"
-            android:text="Tên bài hát hiển thị ở đây"/>
-        <TextView
-            android:id="@+id/txt_singer"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Tên ca sỹ"
-            android:textAlignment="center"/>
-    </LinearLayout>
+            android:layout_height="match_parent"
+            android:text="@string/public_fragment"
+            android:gravity="center" />
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/song_infor"
-        android:gravity="center_horizontal"
-        android:orientation="horizontal">
-        <Button
-            android:id="@+id/btn_start_service"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Start"/>
-        <Button
-            android:id="@+id/btn_stop_service"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Stop"
-            android:layout_marginLeft="10dp"/>
-    </LinearLayout>
+    </FrameLayout>
+<!--<RelativeLayout-->
+<!--    android:layout_width="match_parent"-->
+<!--    android:layout_height="match_parent"-->
+<!--    tools:context=".activities.SecondActivity">-->
+
+<!--    <ImageView-->
+<!--        android:id="@+id/img_song"-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="200dp"-->
+<!--        android:src="@drawable/hutao_banner"-->
+<!--        />-->
+<!--    <LinearLayout-->
+<!--        android:id="@+id/song_infor"-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="100dp"-->
+<!--        android:layout_below="@+id/img_song"-->
+<!--        android:orientation="vertical">-->
+<!--        <TextView-->
+<!--            android:id="@+id/txt_song_name"-->
+<!--            android:layout_width="match_parent"-->
+<!--            android:layout_height="wrap_content"-->
+<!--            android:textAlignment="center"-->
+<!--            android:text="Tên bài hát hiển thị ở đây"/>-->
+<!--        <TextView-->
+<!--            android:id="@+id/txt_singer"-->
+<!--            android:layout_width="match_parent"-->
+<!--            android:layout_height="wrap_content"-->
+<!--            android:text="Tên ca sỹ"-->
+<!--            android:textAlignment="center"/>-->
+<!--    </LinearLayout>-->
+
+<!--    <LinearLayout-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="wrap_content"-->
+<!--        android:layout_below="@+id/song_infor"-->
+<!--        android:gravity="center_horizontal"-->
+<!--        android:orientation="horizontal">-->
+<!--        <Button-->
+<!--            android:id="@+id/btn_start_service"-->
+<!--            android:layout_width="wrap_content"-->
+<!--            android:layout_height="wrap_content"-->
+<!--            android:text="Start"/>-->
+<!--        <Button-->
+<!--            android:id="@+id/btn_stop_service"-->
+<!--            android:layout_width="wrap_content"-->
+<!--            android:layout_height="wrap_content"-->
+<!--            android:text="Stop"-->
+<!--            android:layout_marginLeft="10dp"/>-->
+<!--    </LinearLayout>-->
 
 
 
-</RelativeLayout>
+<!--</RelativeLayout>-->
 </layout>
